// apps/self-serve/src/app/modules/portfolio-recommendation/models/fund-list.model.ts
export interface FundList {
    maximum_initial_subscription_amount_str: string;
    close_date: string;
    three_month_ind: string;
    switch_indicator: string;
    fund_document: any;
    fund_code: string;
    risk_rating: string;
    minimum_initial_subscription_amount: number;
    maximum_subsequent_subscription_amount: number;
    manager_code: string;
    classHexa: string;
    minimum_initial_subscription_amount_str: string;
    class_name: string;
    min_holding: number;
    minimum_subsequent_subscription_amount_str: string;
    current_holding: string;
    min_redem_amt: number;
    fund_status: string;
    class_seq: number | string;
    maximum_initial_subscription_amount: number;
    fund_name: string;
    one_month_ind: string;
    fund_indicator: string;
    esg_fund: string;
    recommended: number;
    manager_name: string;
    three_month: string;
    one_month: string;
    max_redem_amt: number;
    fund_risk_rating: string;
    maximum_subsequent_subscription_amount_str: string;
    nav_price: number;
    min_switch_amt: number;
    risk_name: string;
    product_category: string;
    minimum_subsequent_subscription_amount: number;
    totalInvestment: number;
}